Moving ES-specifics from `MessagesIT` to `MessagesES6IT`.
